calendarapp/requirements.txt
asgiref==3.8.1
click==8.1.7
dj-rest-auth==6.0.0
Django==5.0.7
djangorestframework==3.15.2
djangorestframework-oauth==1.1.0
djangorestframework-simplejwt==5.3.1
drf-yasg==1.21.7
inflection==0.5.1
packaging==24.1
pillow==10.4.0
PyJWT==2.8.0
pytz==2024.1
PyYAML==6.0.1
setuptools==71.1.0
sqlparse==0.5.1
svgwrite==1.4.3
Tree==0.2.4
uritemplate==4.1.1